% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomdata.R
\name{alloneyr}
\alias{alloneyr}
\title{All One Year Function}
\usage{
alloneyr(inputyears, inputformat = "\%Y-\%m-\%d")
}
\arguments{
\item{inputyears}{A vector of the years to be changed}

\item{inputformat}{The format of the inputyears vector. Defaults to "%Y-%m-%d"}
}
\description{
Translates a vector of dates (with months, days and years) into a vector of months-days with no year
}
\examples{
alloneyr()
}
\keyword{date}
\keyword{dates}
