% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomdata.R
\name{testfunction}
\alias{testfunction}
\title{Test Function}
\usage{
testfunction(number)
}
\arguments{
\item{number}{Any integer. It should come out multiplied by 7}
}
\description{
A function to test if this package has been installed correctly
}
\examples{
testfunction()
}
\keyword{function}
\keyword{test}
